Migrate to Java 21
