+++ README.md
@@ -87,7 +87,7 @@ This repository only contains the source code for the package.
 
 ### Set up the prerequisites
 
-1. Download and install Java SE Development Kit (JDK) version 17 (from one of the following locations).
+1. Download and install Java SE Development Kit (JDK) version 21 (from one of the following locations).
     * [Oracle](https://www.oracle.com/java/technologies/downloads/)
 
     * [OpenJDK](https://adoptium.net/)
