+++ ballerina/CompilerPlugin.toml
@@ -3,4 +3,4 @@ id = "constraint-compiler-plugin"
 class = "io.ballerina.lib.data.yaml.compiler.YamlDataCompilerPlugin"
 
 [[dependency]]
-path = "../compiler-plugin/build/libs/data.yaml-compiler-plugin-0.1.0.jar"
+path = "../compiler-plugin/build/libs/data.yaml-compiler-plugin-0.1.1-SNAPSHOT.jar"
