+++ build-config/resources/Ballerina.toml
@@ -11,14 +11,14 @@ distribution = "2201.10.0"
 [platform.java17]
 graalvmCompatible = true
 
-[[platform.java17.dependency]]
+[[platform.java21.dependency]]
 groupId = "io.ballerina.lib"
 artifactId = "yaml-native"
 version = "@toml.version@"
 path = "../native/build/libs/data.yaml-native-@project.version@.jar"
 
 
-[[platform.java17.dependency]]
+[[platform.java21.dependency]]
 groupId = "io.ballerina.stdlib"
 artifactId = "constraint-native"
 version = "@constraint.version@"
