+++ gradle.properties
@@ -1,13 +1,13 @@
 org.gradle.caching=true
 group=io.ballerina.lib
 version=0.1.1-SNAPSHOT
-ballerinaLangVersion=2201.10.0
+ballerinaLangVersion=2201.10.0-20241002-115300-57eee14c
 
 checkstyleToolVersion=10.12.0
 puppycrawlCheckstyleVersion=10.12.0
 testngVersion=7.6.1
 slf4jVersion=2.0.7
-githubSpotbugsVersion=5.0.14
+githubSpotbugsVersion=6.0.18
 githubJohnrengelmanShadowVersion=8.1.1
 underCouchDownloadVersion=4.0.4
 researchgateReleaseVersion=2.8.0
@@ -17,6 +17,6 @@ downloadPluginVersion=4.0.4
 releasePluginVersion=2.8.0
 ballerinaGradlePluginVersion=2.0.1
 
-stdlibIoVersion=1.6.1
-stdlibConstraintVersion=1.5.0
-stdlibTimeVersion=2.4.0
+stdlibIoVersion=1.6.2-20240928-084100-656404f
+stdlibConstraintVersion=1.5.1-20240930-123400-5ecd396
+stdlibTimeVersion=2.5.1-20240930-120200-e59222b
