+++ native/src/main/java/io/ballerina/lib/data/yaml/Native.java
@@ -19,13 +19,12 @@
 package io.ballerina.lib.data.yaml;
 
 import io.ballerina.lib.data.yaml.emitter.Emitter;
-import io.ballerina.lib.data.yaml.io.DataReaderTask;
-import io.ballerina.lib.data.yaml.io.DataReaderThreadPool;
+import io.ballerina.lib.data.yaml.io.BallerinaByteBlockInputStream;
 import io.ballerina.lib.data.yaml.parser.YamlParser;
 import io.ballerina.lib.data.yaml.serializer.Serializer;
+import io.ballerina.lib.data.yaml.utils.DiagnosticLog;
 import io.ballerina.lib.data.yaml.utils.OptionsUtils;
 import io.ballerina.runtime.api.Environment;
-import io.ballerina.runtime.api.Future;
 import io.ballerina.runtime.api.creators.ValueCreator;
 import io.ballerina.runtime.api.values.BArray;
 import io.ballerina.runtime.api.values.BError;
@@ -39,6 +38,9 @@ import java.io.ByteArrayInputStream;
 import java.io.InputStreamReader;
 import java.io.StringReader;
 import java.util.List;
+
+import static io.ballerina.lib.data.yaml.utils.DataReader.resolveCloseMethod;
+import static io.ballerina.lib.data.yaml.utils.DataReader.resolveNextMethod;
 
 /**
  * This class is used to convert json inform of string, byte[], byte-stream to record or json type.
@@ -69,10 +71,17 @@ public class Native {
 
     public static Object parseStream(Environment env, BStream yaml, BMap<BString, Object> options, BTypedesc typed) {
         final BObject iteratorObj = yaml.getIteratorObj();
-        final Future future = env.markAsync();
-        DataReaderTask task = new DataReaderTask(env, iteratorObj, future, typed, options);
-        DataReaderThreadPool.EXECUTOR_SERVICE.submit(task);
-        return null;
+        try {
+            BallerinaByteBlockInputStream byteBlockSteam = new BallerinaByteBlockInputStream(env,
+                    iteratorObj, resolveNextMethod(iteratorObj), resolveCloseMethod(iteratorObj));
+            Object result = YamlParser.compose(new InputStreamReader(byteBlockSteam), options, typed);
+            if (byteBlockSteam.getError() != null) {
+                return byteBlockSteam.getError();
+            }
+            return result;
+        } catch (Exception e) {
+            return DiagnosticLog.getYamlError("Error occurred while reading the stream: " + e.getMessage());
+        }
     }
 
     public static Object toYamlStringArray(Object yamlValue, BMap<BString, Object> config) {
